+++ navi/navi.php
@@ -30,11 +30,11 @@ class Navi extends Module
     const DEFAULT_FAB_POSITION = 'right';
     const REPO_URL = 'https://github.com/Lucas-tsl/navi';
 
-    const DEFAULT_STORIES_VIDEO_BORDER_WIDTH = '0';
+    const DEFAULT_STORIES_PHONE_PADDING = '10';
     const DEFAULT_STORIES_PHONE_WIDTH = '200';
     const MIN_STORIES_PHONE_WIDTH = 150;
     const MAX_STORIES_PHONE_WIDTH = 280;
-    const MAX_STORIES_VIDEO_BORDER_WIDTH = 10;
+    const MAX_STORIES_PHONE_PADDING = 20;
 
     /**
      * Bascules "Afficher sur ordinateur / mobile" partagées par toutes les
@@ -68,7 +68,7 @@ class Navi extends Module
     {
         $this->name = 'navi';
         $this->tab = 'front_office_features';
-        $this->version = '1.5.0';
+        $this->version = '1.5.1';
         $this->author = 'Troteseil Lucas';
         $this->need_instance = 0;
         $this->bootstrap = true;
@@ -171,7 +171,7 @@ class Navi extends Module
         Configuration::updateValue('NAVI_STORIES_COLOR_CLOSE_BG', self::DEFAULT_STORIES_CLOSE_BG);
         Configuration::updateValue('NAVI_STORIES_COLOR_OVERLAY', self::DEFAULT_STORIES_OVERLAY_BG);
         Configuration::updateValue('NAVI_FAB_POSITION', self::DEFAULT_FAB_POSITION);
-        Configuration::updateValue('NAVI_STORIES_VIDEO_BORDER_WIDTH', self::DEFAULT_STORIES_VIDEO_BORDER_WIDTH);
+        Configuration::updateValue('NAVI_STORIES_PHONE_PADDING', self::DEFAULT_STORIES_PHONE_PADDING);
         Configuration::updateValue('NAVI_STORIES_PHONE_WIDTH', self::DEFAULT_STORIES_PHONE_WIDTH);
 
         foreach (self::VISIBILITY_TOGGLES as $desktopKey => $info) {
@@ -227,7 +227,7 @@ class Navi extends Module
             && Configuration::deleteByName('NAVI_STORIES_COLOR_CLOSE_BG')
             && Configuration::deleteByName('NAVI_STORIES_COLOR_OVERLAY')
             && Configuration::deleteByName('NAVI_FAB_POSITION')
-            && Configuration::deleteByName('NAVI_STORIES_VIDEO_BORDER_WIDTH')
+            && Configuration::deleteByName('NAVI_STORIES_PHONE_PADDING')
             && Configuration::deleteByName('NAVI_STORIES_PHONE_WIDTH');
 
         foreach (self::VISIBILITY_TOGGLES as $desktopKey => $info) {
@@ -330,9 +330,9 @@ class Navi extends Module
         }
 
         if (Tools::isSubmit('submitNaviVideoAppearance')) {
-            $borderWidth = max(0, min(self::MAX_STORIES_VIDEO_BORDER_WIDTH, (int) Tools::getValue('NAVI_STORIES_VIDEO_BORDER_WIDTH')));
+            $phonePadding = max(0, min(self::MAX_STORIES_PHONE_PADDING, (int) Tools::getValue('NAVI_STORIES_PHONE_PADDING')));
             $phoneWidth = max(self::MIN_STORIES_PHONE_WIDTH, min(self::MAX_STORIES_PHONE_WIDTH, (int) Tools::getValue('NAVI_STORIES_PHONE_WIDTH')));
-            Configuration::updateValue('NAVI_STORIES_VIDEO_BORDER_WIDTH', $borderWidth);
+            Configuration::updateValue('NAVI_STORIES_PHONE_PADDING', $phonePadding);
             Configuration::updateValue('NAVI_STORIES_PHONE_WIDTH', $phoneWidth);
             $output .= $this->displayConfirmation($this->l('Réglages enregistrés.'));
         }
@@ -352,8 +352,8 @@ class Navi extends Module
     private function getVideoAppearanceBlock()
     {
         $this->context->smarty->assign([
-            'navi_video_border_width' => max(0, (int) (Configuration::get('NAVI_STORIES_VIDEO_BORDER_WIDTH') !== false ? Configuration::get('NAVI_STORIES_VIDEO_BORDER_WIDTH') : self::DEFAULT_STORIES_VIDEO_BORDER_WIDTH)),
-            'navi_video_border_max' => self::MAX_STORIES_VIDEO_BORDER_WIDTH,
+            'navi_phone_padding' => max(0, (int) (Configuration::get('NAVI_STORIES_PHONE_PADDING') !== false ? Configuration::get('NAVI_STORIES_PHONE_PADDING') : self::DEFAULT_STORIES_PHONE_PADDING)),
+            'navi_phone_padding_max' => self::MAX_STORIES_PHONE_PADDING,
             'navi_phone_width' => max(self::MIN_STORIES_PHONE_WIDTH, (int) (Configuration::get('NAVI_STORIES_PHONE_WIDTH') !== false ? Configuration::get('NAVI_STORIES_PHONE_WIDTH') : self::DEFAULT_STORIES_PHONE_WIDTH)),
             'navi_phone_width_min' => self::MIN_STORIES_PHONE_WIDTH,
             'navi_phone_width_max' => self::MAX_STORIES_PHONE_WIDTH,
@@ -361,7 +361,7 @@ class Navi extends Module
             'navi_ajax_token' => Tools::getAdminTokenLite('AdminModules'),
             'navi_current_index' => $this->context->link->getAdminLink('AdminModules', false)
                 . '&configure=' . $this->name . '&tab_module=' . $this->tab . '&module_name=' . $this->name,
-            'navi_video_border_ticks' => range(0, self::MAX_STORIES_VIDEO_BORDER_WIDTH),
+            'navi_phone_padding_ticks' => range(0, self::MAX_STORIES_PHONE_PADDING, 2),
             'navi_phone_width_ticks' => range(self::MIN_STORIES_PHONE_WIDTH, self::MAX_STORIES_PHONE_WIDTH, 10),
         ]);
 
@@ -736,7 +736,7 @@ class Navi extends Module
         $storiesCloseIcon = $this->sanitizeHexColor(Configuration::get('NAVI_STORIES_COLOR_CLOSE_ICON'), self::DEFAULT_STORIES_CLOSE_ICON);
         $storiesCloseBg = $this->sanitizeHexColor(Configuration::get('NAVI_STORIES_COLOR_CLOSE_BG'), self::DEFAULT_STORIES_CLOSE_BG);
         $storiesOverlayBg = $this->sanitizeHexColor(Configuration::get('NAVI_STORIES_COLOR_OVERLAY'), self::DEFAULT_STORIES_OVERLAY_BG);
-        $videoBorderWidth = max(0, min(self::MAX_STORIES_VIDEO_BORDER_WIDTH, (int) (Configuration::get('NAVI_STORIES_VIDEO_BORDER_WIDTH') !== false ? Configuration::get('NAVI_STORIES_VIDEO_BORDER_WIDTH') : self::DEFAULT_STORIES_VIDEO_BORDER_WIDTH)));
+        $phonePadding = max(0, min(self::MAX_STORIES_PHONE_PADDING, (int) (Configuration::get('NAVI_STORIES_PHONE_PADDING') !== false ? Configuration::get('NAVI_STORIES_PHONE_PADDING') : self::DEFAULT_STORIES_PHONE_PADDING)));
         $phoneWidth = max(self::MIN_STORIES_PHONE_WIDTH, min(self::MAX_STORIES_PHONE_WIDTH, (int) (Configuration::get('NAVI_STORIES_PHONE_WIDTH') !== false ? Configuration::get('NAVI_STORIES_PHONE_WIDTH') : self::DEFAULT_STORIES_PHONE_WIDTH)));
 
         $css = 'html:root{'
@@ -749,7 +749,7 @@ class Navi extends Module
             . '--navi-story-close-icon:' . $storiesCloseIcon . ';'
             . '--navi-story-close-bg:' . $storiesCloseBg . ';'
             . '--navi-story-overlay-bg:' . $storiesOverlayBg . ';'
-            . '--navi-story-video-border-width:' . $videoBorderWidth . 'px;'
+            . '--navi-story-phone-padding:' . $phonePadding . 'px;'
             . '--navi-story-phone-width:' . $phoneWidth . 'px;'
             . '}';